Move class definitions to header file
